feat: node-sourcer bundle dependencies into final dist by default (#18)
I'm expecting that all things which are placed in `dependencies` of a
node package or express app - will be bundled into the final code.

Exept cases, when a package is added to `peerDependencies` - in that
case we don't need to add `node_modules` into `dist ` code, and we need
to add a package in a final app + bundle it with a final bundler.